split some components into subcomponents
